+++ src/main/java/CoffeeOrderWebApp/practice/coffee/service/CoffeeService.java
@@ -4,27 +4,30 @@ import CoffeeOrderWebApp.practice.coffee.entity.Coffee;
 import CoffeeOrderWebApp.practice.coffee.repository.CoffeeRepository;
 import CoffeeOrderWebApp.practice.exception.ExceptionEnum;
 import CoffeeOrderWebApp.practice.exception.LogicException;
+import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
+import javax.transaction.Transactional;
 import java.util.Optional;
 
 @Service
+@RequiredArgsConstructor
+@Transactional
 public class CoffeeService {
-    private CoffeeRepository coffeeRepository;
-
-    public CoffeeService(CoffeeRepository coffeeRepository) {
-        this.coffeeRepository = coffeeRepository;
-    }
+    private final CoffeeRepository coffeeRepository;
 
     public Coffee createCoffee(Coffee coffee){
+        verifyAdmin();
         verifyExistCoffee(coffee);
         return coffeeRepository.save(coffee);
     }
 
     public Coffee modifyCoffee(Coffee coffee){
+        verifyAdmin();
         Coffee foundCoffee = findCoffee(coffee.getCoffeeId());
 
         Optional.ofNullable(coffee.getKorName()).ifPresent(name -> foundCoffee.setKorName(name));
@@ -45,6 +48,7 @@ public class CoffeeService {
     }
 
     public void deleteCoffee(long coffeeId){
+        verifyAdmin();
         findCoffee(coffeeId);
         coffeeRepository.deleteById(coffeeId);
     }
@@ -58,5 +62,10 @@ public class CoffeeService {
         Optional<Coffee> optionalCoffeeEntity = coffeeRepository.findById(coffeeId);
         Coffee foundCoffee = optionalCoffeeEntity.orElseThrow(()->new LogicException(ExceptionEnum.COFFEE_NOT_FOUND));
         return foundCoffee;
+    }
+
+    private void verifyAdmin() {
+        String principal = SecurityContextHolder.getContext().getAuthentication().getAuthorities().toString();
+        if(!principal.contains("ROLE_ADMIN")) throw new LogicException(ExceptionEnum.ADMIN_ACCESS_ONLY);;
     }
 }
